organize code a bit better
