+++ lib/requirements.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require           "erb"
+require           "json"
+require           "pastel"
+require           "psych"
+require           "rest-client"
+require           "thor"
+require           "tty-config"
+require           "tty-editor"
+require           "tty-font"
+require           "tty-prompt"
+require           "tty-spinner"
+require           "tty-table"
+require_relative  "rockette/cli"
+require_relative  "rockette/controller"
+require_relative  "rockette/rester"
+require_relative  "rockette/text_helper"
+require_relative  "rockette/version"
